This adds the ability to set consumer group offsets
I added several other methods to the API that should improve quality of
life.

New methods:

* GetTopicPartitions(ctx context.Context, topic string) ([]int, error)
* GetTopicOffsets(ctx context.Context, topic string) ([]kafka.PartitionOffsets, error)
* CreateConsumerGroup(ctx context.Context, topic, cg string) (*kafka.ConsumerGroup, error)
* DeleteConsumerGroup(ctx context.Context, cg string) error
* GetConsumerGroupOffsets(ctx context.Context, topic, cg string) (map[int]int64, error)
* SetConsumerGroupOffsets(ctx context.Context, topic, cg string, offsets map[int]int64) error

Set and Get consumer being the most important ones.

What we were going for was being able to "clone" consumer group settings
from an existing consumer group and applying those settings in another
consumer group.

To do that you would first do `GetConsumerGroupOffsets()` on the
original cg; then run `SetConsumerGroupOffsets()` for the new cg and
pass in the offsets that you received from `Get...()`.

This is using a combination of sarama and segment because both have only
partial or working functionality for pulling this off.

Sadly, sarama's `ResetOffsets()` via OffsetManager does not work.

:phew: